Networking: Remove eth0 from bridge br0 and enable DHCP on eth0
We remove eth0 from br0 and enable DHCP because we connect the PI via
eth0 to the router/accesspoint that is part of the antenna box. Or we
connect the PI directly to the board network of the ship via the
eth0 port. In this case we disable the router or configure the router as
bridge/switch and dhcp client.