Send nmea position information from gpsd to GCS clients
The gps2udp service will send nmea message to the GCS client via UDP.
You have to add the IP/port of the client to the list of clients with
'-u'. E.g. in QGC unstable you can enable listening for nmea message in
the application settings. This will allow devices without GPS to see
the GCS/ship/home position.