Implement ADS-B receiver with HTML frontend
This adds an software defined ADS-B receiver based on readsb to the GCS.
We use tar1090 to display detected airplanes. Ngnix serves the static
HTML fronted to the user via port 80. You can use QGC to connect to the
ADS-B receiver and visualize approaching airplanes in QGC. Please use
the GCS hostname/IP and the 30003 port for that. The ADS-B receiver
connects to GPSD as well to get position information. The frontend will
update the position and e.g. the receiver position and distance to the
planes will move with the ship.