Add SRTM Downloading for offline usage
SRTM tiles contain information about terrain altitude. We use the
srtm_downloader.py in the CI build to download SRTM tiles for offline
use. Later on the MAVProxy moving home module uses the tiles to get the
altitude for the position of the GCS. At sea level the altitude is
always 0 therefore we do not download tiles e.g. for the Med.

Please specify your airfields in a GeoJSON file and add it to the list
of areas to download.

We need altitude information for autolanding purposes.