Add mav-ping tool
This adds a rudimentary tool to estimate the round trip time of a
MAVLink based communication system. Based on the time of arrival we can
observe packet loss as well. To connect to the local mavlink-router
instanace that is connected to the UAV use something along the lines of:

        mav-ping tcp:127.0.0.1:5760

I will try to add a more sophisticated version that highlights packet
loss or late arrivals better in the future.